# pkg/api/v1/namespace_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.32.1
// source: api/v1/namespace.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationNamespaceCreateNamespace = "/sovereign.api.v1.Namespace/CreateNamespace"
const OperationNamespaceDeleteNamespace = "/sovereign.api.v1.Namespace/DeleteNamespace"
const OperationNamespaceGetNamespace = "/sovereign.api.v1.Namespace/GetNamespace"
const OperationNamespaceListNamespace = "/sovereign.api.v1.Namespace/ListNamespace"
const OperationNamespaceSelectNamespace = "/sovereign.api.v1.Namespace/SelectNamespace"
const OperationNamespaceUpdateNamespace = "/sovereign.api.v1.Namespace/UpdateNamespace"
const OperationNamespaceUpdateNamespaceStatus = "/sovereign.api.v1.Namespace/UpdateNamespaceStatus"

type NamespaceHTTPServer interface {
	CreateNamespace(context.Context, *CreateNamespaceRequest) (*CreateNamespaceReply, error)
	DeleteNamespace(context.Context, *DeleteNamespaceRequest) (*DeleteNamespaceReply, error)
	GetNamespace(context.Context, *GetNamespaceRequest) (*NamespaceItem, error)
	ListNamespace(context.Context, *ListNamespaceRequest) (*ListNamespaceReply, error)
	SelectNamespace(context.Context, *SelectNamespaceRequest) (*SelectNamespaceReply, error)
	UpdateNamespace(context.Context, *UpdateNamespaceRequest) (*UpdateNamespaceReply, error)
	UpdateNamespaceStatus(context.Context, *UpdateNamespaceStatusRequest) (*UpdateNamespaceStatusReply, error)
}

func RegisterNamespaceHTTPServer(s *http.Server, srv NamespaceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/namespace", _Namespace_CreateNamespace0_HTTP_Handler(srv))
	r.PUT("/v1/namespace/{uid}", _Namespace_UpdateNamespace0_HTTP_Handler(srv))
	r.PUT("/v1/namespace/{uid}/status", _Namespace_UpdateNamespaceStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/namespace/{uid}", _Namespace_DeleteNamespace0_HTTP_Handler(srv))
	r.GET("/v1/namespace/{uid}", _Namespace_GetNamespace0_HTTP_Handler(srv))
	r.GET("/v1/namespaces", _Namespace_ListNamespace0_HTTP_Handler(srv))
	r.GET("/v1/namespaces/select", _Namespace_SelectNamespace0_HTTP_Handler(srv))
}

func _Namespace_CreateNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateNamespaceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceCreateNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateNamespace(ctx, req.(*CreateNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateNamespaceReply)
		return ctx.Result(200, reply)
	}
}

func _Namespace_UpdateNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateNamespaceRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceUpdateNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateNamespace(ctx, req.(*UpdateNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateNamespaceReply)
		return ctx.Result(200, reply)
	}
}

func _Namespace_UpdateNamespaceStatus0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateNamespaceStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceUpdateNamespaceStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateNamespaceStatus(ctx, req.(*UpdateNamespaceStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateNamespaceStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Namespace_DeleteNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteNamespaceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceDeleteNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteNamespace(ctx, req.(*DeleteNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteNamespaceReply)
		return ctx.Result(200, reply)
	}
}

func _Namespace_GetNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetNamespaceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceGetNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetNamespace(ctx, req.(*GetNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*NamespaceItem)
		return ctx.Result(200, reply)
	}
}

func _Namespace_ListNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListNamespaceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceListNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListNamespace(ctx, req.(*ListNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListNamespaceReply)
		return ctx.Result(200, reply)
	}
}

func _Namespace_SelectNamespace0_HTTP_Handler(srv NamespaceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectNamespaceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNamespaceSelectNamespace)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectNamespace(ctx, req.(*SelectNamespaceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectNamespaceReply)
		return ctx.Result(200, reply)
	}
}

type NamespaceHTTPClient interface {
	CreateNamespace(ctx context.Context, req *CreateNamespaceRequest, opts ...http.CallOption) (rsp *CreateNamespaceReply, err error)
	DeleteNamespace(ctx context.Context, req *DeleteNamespaceRequest, opts ...http.CallOption) (rsp *DeleteNamespaceReply, err error)
	GetNamespace(ctx context.Context, req *GetNamespaceRequest, opts ...http.CallOption) (rsp *NamespaceItem, err error)
	ListNamespace(ctx context.Context, req *ListNamespaceRequest, opts ...http.CallOption) (rsp *ListNamespaceReply, err error)
	SelectNamespace(ctx context.Context, req *SelectNamespaceRequest, opts ...http.CallOption) (rsp *SelectNamespaceReply, err error)
	UpdateNamespace(ctx context.Context, req *UpdateNamespaceRequest, opts ...http.CallOption) (rsp *UpdateNamespaceReply, err error)
	UpdateNamespaceStatus(ctx context.Context, req *UpdateNamespaceStatusRequest, opts ...http.CallOption) (rsp *UpdateNamespaceStatusReply, err error)
}

type NamespaceHTTPClientImpl struct {
	cc *http.Client
}

func NewNamespaceHTTPClient(client *http.Client) NamespaceHTTPClient {
	return &NamespaceHTTPClientImpl{client}
}

func (c *NamespaceHTTPClientImpl) CreateNamespace(ctx context.Context, in *CreateNamespaceRequest, opts ...http.CallOption) (*CreateNamespaceReply, error) {
	var out CreateNamespaceReply
	pattern := "/v1/namespace"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNamespaceCreateNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) DeleteNamespace(ctx context.Context, in *DeleteNamespaceRequest, opts ...http.CallOption) (*DeleteNamespaceReply, error) {
	var out DeleteNamespaceReply
	pattern := "/v1/namespace/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNamespaceDeleteNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) GetNamespace(ctx context.Context, in *GetNamespaceRequest, opts ...http.CallOption) (*NamespaceItem, error) {
	var out NamespaceItem
	pattern := "/v1/namespace/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNamespaceGetNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) ListNamespace(ctx context.Context, in *ListNamespaceRequest, opts ...http.CallOption) (*ListNamespaceReply, error) {
	var out ListNamespaceReply
	pattern := "/v1/namespaces"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNamespaceListNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) SelectNamespace(ctx context.Context, in *SelectNamespaceRequest, opts ...http.CallOption) (*SelectNamespaceReply, error) {
	var out SelectNamespaceReply
	pattern := "/v1/namespaces/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNamespaceSelectNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) UpdateNamespace(ctx context.Context, in *UpdateNamespaceRequest, opts ...http.CallOption) (*UpdateNamespaceReply, error) {
	var out UpdateNamespaceReply
	pattern := "/v1/namespace/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNamespaceUpdateNamespace))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NamespaceHTTPClientImpl) UpdateNamespaceStatus(ctx context.Context, in *UpdateNamespaceStatusRequest, opts ...http.CallOption) (*UpdateNamespaceStatusReply, error) {
	var out UpdateNamespaceStatusReply
	pattern := "/v1/namespace/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNamespaceUpdateNamespaceStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}